rename deletion_migration -> deletion_migrations
